requirements.txt
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2
Django==5.1
django-cors-headers==3.13.0
django-filter==24.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.7.1
python-dotenv==0.21.0
redis==5.0.8